+++ scout/commands/export/export_command.py
@@ -9,7 +9,7 @@ Created by Måns Magnusson on 2016-05-11.
 Copyright (c) 2016 ScoutTeam__. All rights reserved.
 
 """
-
+import copy
 import logging
 
 import click
@@ -40,7 +40,15 @@ export.add_command(panel_cmd)
 export.add_command(genes)
 export.add_command(transcripts)
 export.add_command(exons)
+# Register normally
 export.add_command(causatives)
+
+# Create a shallow copy of the command for the alias
+variants_cmd = copy.copy(causatives)
+variants_cmd.hidden = True
+
+# Register the alias
+export.add_command(variants_cmd, "variants")
 export.add_command(verified)
 export.add_command(managed)
 export.add_command(hpo_genes)